feat(dx): Move dashboard endpoints to dashboard module
