+++ tests/sentry/dashboards/endpoints/test_organization_dashboard_details.py
