+++ tests/sentry/dashboards/endpoints/test_organization_dashboard_widget_details.py
