+++ tests/sentry/dashboards/endpoints/test_organization_dashboards_starred.py
